feat(gatsby-plugin-clerk): Remove interstitial
