+++ packages/gatsby-plugin-clerk/src/ssr/utils.ts
@@ -3,6 +3,9 @@ import { prunePrivateMetadata } from '@clerk/backend';
 import cookie from 'cookie';
 import type { GetServerDataProps } from 'gatsby';
 
+import { SECRET_KEY } from '../constants';
+import { constants } from './clerkClient';
+
 /**
  * @internal
  */
@@ -38,19 +41,43 @@ export const wrapWithClerkState = (data: any) => {
   return { clerkState: { __internal_clerk_state: { ...data } } };
 };
 
-/**
- * @internal
- */
 export const parseCookies = (headers: any) => {
   return cookie.parse(headers.get('cookie') || '');
 };
 
-/**
- * @internal
- */
 export function injectSSRStateIntoProps(callbackResult: any, data: any) {
   return {
     ...callbackResult,
     props: { ...callbackResult.props, ...wrapWithClerkState(data) },
   };
+}
+
+export const gatsbyPropsToRequest = (context: GetServerDataProps): Request => {
+  const headers = new Headers(Object.fromEntries(context.headers) as Record<string, string>);
+  headers.set(constants.Headers.ForwardedHost, returnReferrerAsXForwardedHostToFixLocalDevGatsbyProxy(context.headers));
+  return new Request(context.url, { method: context.method, headers });
+};
+
+const returnReferrerAsXForwardedHostToFixLocalDevGatsbyProxy = (headers: Map<string, unknown>) => {
+  if (process.env.NODE_ENV !== 'development') {
+    return headers.get(constants.Headers.ForwardedHost) as string;
+  }
+
+  const forwardedHost = headers.get(constants.Headers.ForwardedHost) as string;
+  if (forwardedHost) {
+    return forwardedHost;
+  }
+
+  const referrerUrl = new URL(headers.get(constants.Headers.Referrer) as string);
+  const hostUrl = new URL('https://' + (headers.get(constants.Headers.Host) as string));
+
+  if (isDevelopmentOrStaging(SECRET_KEY || '') && hostUrl.hostname === referrerUrl.hostname) {
+    return referrerUrl.host;
+  }
+
+  return forwardedHost;
+};
+
+function isDevelopmentOrStaging(apiKey: string): boolean {
+  return apiKey.startsWith('test_') || apiKey.startsWith('sk_test_');
 }
